Backend/requirement.txt
Flask==2.1.3
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask_Cors=3.0.10